# perf(merkle): optimize hashing by reusing sha256 in a pool