@rootulp 2024-03-04
I think this is relevant:

https://github.com/celestiaorg/go-square/blob/4e84d80f3e55ac3093005c8089186d68ee62541d/merkle/README.md?plain=1#L6

I suppose we could also copy + paste [this license](https://github.com/cometbft/cometbft/blob/main/LICENSE) into that directory.