@cmwaters 2024-06-28
Closing this because for now we've decided to keep the merkle library in our cometbft fork. Thanks for opening the PR in any case. Sorry that we didn't merge it. We may look to apply this to our cometbft fork in the future